feat: open/close edit form
