--- src/view/edit/header/delete-button-view.js
@@ -1,12 +1,0 @@
-import AbstractView from '/src/framework/view/abstract-view';
-
-
-const createDeleteButtonTemplate = () =>
-  '<button class="event__reset-btn" type="reset">Delete</button>';
-
-
-export default class DeleteButtonView extends AbstractView {
-  get template() {
-    return createDeleteButtonTemplate();
-  }
-}
