--- src/view/edit/header/rollup-button-view.js
@@ -1,14 +1,0 @@
-import AbstractView from '/src/framework/view/abstract-view';
-
-
-const createRollupButtonTemplate = () =>
-  `<button class="event__rollup-btn" type="button">
-    <span class="visually-hidden">Open event</span>
-  </button>`;
-
-
-export default class RollupButtonView extends AbstractView {
-  get template() {
-    return createRollupButtonTemplate();
-  }
-}
